+++ src/BinaryFileHandler/md5.h
@@ -1,0 +1,8 @@
+#ifndef WIIMAKE_MD5_H
+#define WIIMAKE_MD5_H
+
+#include <string>
+
+std::string md5(std::string);
+
+#endif