+++ src/MainProgram/wiimake-isotool.cpp
@@ -1,6 +1,7 @@
-#include "../ArgumentParsing/Parser.h"
-#include "../ISO.h"
-#include "../Global.h"
+#include "Parser.h"
+#include "ISO.h"
+#include "md5.h"
+#include "Global.h"
 #include "Description.h"
 
 #include <iostream>
@@ -14,7 +15,6 @@ void run(TokenList& tokens)
     if (tokens[1] == "--read")
     {
         INVALID_ARG(tokens.size() != 3, "incorrect number of arguments");
-        /* make sure valid address is given */
         try
         {
             std::cout << std::hex << iso.read(tokens.back()) << std::endl;
@@ -27,10 +27,6 @@ void run(TokenList& tokens)
             throw e;
         }
     }
-    else if (tokens[1] == "--checksum")
-    {
-        std::cout << std::hex << iso.checkSum() << std::endl;
-    }
     else if (tokens[1] == "--save")
     {
         INVALID_ARG(tokens.size() != 3, "incorrect number of arguments");
@@ -40,6 +36,22 @@ void run(TokenList& tokens)
     {
         INVALID_ARG(tokens.size() != 3, "incorrect number of arguments");
         iso.loadState(tokens.back());
+    }
+    else if (tokens[1] == "--md5")
+    {
+        std::cout << md5(tokens[0]) << std::endl;
+    }
+    else if (tokens[1] == "--zero")
+    {
+        INVALID_ARG(tokens.size() != 4, "incorrect number of arguments");   
+
+        uint32_t address = stoul(tokens[2], nullptr, 16);
+        uint32_t end = address + stoul(tokens[3], nullptr, 16);
+
+        for (; address < end; address += 0x04)
+        {
+            iso.write(address, 0);
+        }
     }
     else
     {